added cicd workflow
