+++ frontend/votingpreference.html
@@ -1,0 +1,44 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Vote for Your Candidate</title>
+    <link rel="stylesheet" href="vote.css">
+    <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
+</head>
+<body>
+        <div class="container">
+            <h1>Kindly Indicate Your Voting Preference</h1>
+        
+            <!-- Candidate Options -->
+            <form id="voteForm">
+                <label class="candidate-label">
+                    <input type="radio" id="candidate1" name="candidate" value="1" style="display:none;">
+                    <img src="donald trumph.png" alt="Donald Trump" class="candidate-image">
+                    <div>Candidate 1: Donald Trump</div>
+                </label>
+                <label class="candidate-label">
+                    <input type="radio" id="candidate2" name="candidate" value="2" style="display:none;">
+                    <img src="kamala harris.png" alt="Kamala Harris" class="candidate-image">
+                    <div>Candidate 2: Kamala Harris</div>
+                </label>
+            </form>
+        
+            <div class="button-group">
+                <button id="submitButton" class="button-submit" onclick="submitVote()">Submit Vote Preference</button>
+                <button id="exitButton" class="button-exit" onclick="redirectToHome()">Exit</button>
+            </div>
+        </div>
+        
+        <!-- Graphs Section -->
+        <div class="graphs" id="graphs" style="display: none;">
+            <canvas id="ageGraph"></canvas>
+            <canvas id="genderGraph"></canvas>
+            <canvas id="professionGraph"></canvas>
+        </div>
+
+    <script src="vote.js"></script>
+</body>
+</html>
