--- vote.css
@@ -1,127 +1,0 @@
-body {
-    font-family: Arial, sans-serif;
-    background-color: #f4f4f4;
-    text-align: center;
-    padding: 20px;
-}
-
-.container {
-    max-width: 800px;
-    margin: auto;
-    background: #fff;
-    padding: 20px;
-    border-radius: 8px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-}
-
-.candidates {
-    display: flex;
-    justify-content: space-around;
-    margin-bottom: 20px;
-}
-
-.candidate {
-    text-align: center;
-}
-
-.photo {
-    width: 150px;
-    height: 150px;
-    border-radius: 50%;
-    margin-bottom: 10px;
-}
-
-/* Button Styling */
-button {
-    padding: 10px 20px;
-    font-size: 16px;
-    border: none;
-    background-color: #f02626;
-    color: white;
-    border-radius: 5px;
-    cursor: pointer;
-}
-
-button:hover {
-    background-color: #218838;
-}
-
-/* Button Group for Submit and Exit buttons */
-.button-group {
-    display: flex;
-    justify-content: center;
-    gap: 20px; /* Space between the buttons */
-    margin-top: 20px;
-}
-
-.button-submit {
-    background-color: #f02626;
-    color: white;
-    border: none;
-    padding: 10px 20px;
-    font-size: 16px;
-    cursor: pointer;
-    border-radius: 5px;
-}
-
-.button-submit.voted {
-    background-color: #15861b; /* Darker green when voted */
-}
-
-.candidate-label {
-    display: inline-block;
-    cursor: pointer;
-    margin: 50px;
-}
-
-.candidate-image {
-    width: 150px;
-    height: 150px;
-    border-radius: 50%;
-    object-fit: cover;
-    transition: transform 0.3s, border 0.3s;
-    border: 5px solid transparent;
-}
-
-.candidate-label input:checked + .candidate-image {
-    outline: none;
-    border: 5px solid #4CAF50;
-    transition: border 0.3s;
-}
-
-.candidate-image:hover {
-    transform: scale(1.1);
-}
-
-.candidate-label div {
-    margin-top: 5px;
-    font-weight: bold;
-}
-
-/* Graph Styling */
-#graphs {
-    display: none;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    gap: 20px;
-    overflow-x: auto;
-    padding: 10px;
-    border: 1px solid #ccc;
-    border-radius: 8px;
-}
-
-canvas {
-    width: 250px !important;
-    height: 150px !important;
-    flex-shrink: 0;
-    border: 1px solid #ddd;
-    border-radius: 8px;
-}
-
-.button-group {
-    display: flex;
-    justify-content: center;
-    gap: 20px; /* Adjusts space between buttons */
-    margin-top: 20px;
-}
